--- huruhuru/src/main/java/com/hruhuru/huruhuru/controller/MemberController.java
@@ -1,4 +1,0 @@
-package com.hruhuru.huruhuru.controller;
-
-public class MemberController {
-}
