--- huruhuru/src/main/java/com/hruhuru/huruhuru/dto/request/member/MemberSignInRequest.java
@@ -1,4 +1,0 @@
-package com.hruhuru.huruhuru.dto.request.member;
-
-public class MemberSignInRequest {
-}
