--- huruhuru/src/main/java/com/hruhuru/huruhuru/dto/response/member/MemberGetResponse.java
@@ -1,4 +1,0 @@
-package com.hruhuru.huruhuru.dto.response.member;
-
-public class MemberGetResponse {
-}
