--- huruhuru/src/main/java/com/hruhuru/huruhuru/repository/MemberRepository.java
@@ -1,4 +1,0 @@
-package com.hruhuru.huruhuru.repository;
-
-public class MemberRepository {
-}
