--- huruhuru/src/main/java/com/hruhuru/huruhuru/service/MemberService.java
@@ -1,4 +1,0 @@
-package com.hruhuru.huruhuru.service;
-
-public class MemberService {
-}
