+++ huruhuru/src/main/java/com/huruhuru/huruhuru/HuruhuruApplication.java
@@ -1,4 +1,4 @@
-package com.hruhuru.huruhuru;
+package com.huruhuru.huruhuru;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
