+++ huruhuru/src/main/java/com/huruhuru/huruhuru/controller/MemberController.java
@@ -1,0 +1,4 @@
+package com.huruhuru.huruhuru.controller;
+
+public class MemberController {
+}
