+++ huruhuru/src/main/java/com/huruhuru/huruhuru/domain/entity/AnswerEntity.java
@@ -1,8 +1,6 @@
-package com.hruhuru.huruhuru.domain.entity;
+package com.huruhuru.huruhuru.domain.entity;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
 
 @Embeddable
 public class AnswerEntity {
