+++ huruhuru/src/main/java/com/huruhuru/huruhuru/domain/entity/ScoreEntity.java
@@ -1,4 +1,4 @@
-package com.hruhuru.huruhuru.domain.entity;
+package com.huruhuru.huruhuru.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
