+++ huruhuru/src/main/java/com/huruhuru/huruhuru/dto/request/member/MemberSignInRequest.java
@@ -1,0 +1,4 @@
+package com.huruhuru.huruhuru.dto.request.member;
+
+public class MemberSignInRequest {
+}
