+++ huruhuru/src/main/java/com/huruhuru/huruhuru/dto/response/member/MemberGetResponse.java
@@ -1,0 +1,4 @@
+package com.huruhuru.huruhuru.dto.response.member;
+
+public class MemberGetResponse {
+}
