+++ huruhuru/src/main/java/com/huruhuru/huruhuru/global/config/PropertyConfig.java
@@ -1,7 +1,6 @@
-package com.hruhuru.huruhuru.global.config;
+package com.huruhuru.huruhuru.global.config;
 
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 //@PropertySource("classpath:/.env")
