+++ huruhuru/src/main/java/com/huruhuru/huruhuru/repository/MemberRepository.java
@@ -1,0 +1,4 @@
+package com.huruhuru.huruhuru.repository;
+
+public class MemberRepository {
+}
