+++ huruhuru/src/main/java/com/huruhuru/huruhuru/repository/QuestionJpaRepository.java
@@ -1,11 +1,10 @@
-package com.hruhuru.huruhuru.repository;
+package com.huruhuru.huruhuru.repository;
 
-import com.hruhuru.huruhuru.domain.entity.QuestionEntity;
+import com.huruhuru.huruhuru.domain.entity.QuestionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.security.core.parameters.P;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
