+++ huruhuru/src/main/java/com/huruhuru/huruhuru/repository/ScoreJpaRepository.java
@@ -1,0 +1,7 @@
+package com.huruhuru.huruhuru.repository;
+
+import com.huruhuru.huruhuru.domain.entity.ScoreEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ScoreJpaRepository extends JpaRepository<ScoreEntity, Long> {
+}
