+++ huruhuru/src/main/java/com/huruhuru/huruhuru/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.huruhuru.huruhuru.service;
+
+public class MemberService {
+}
